requirements.txt
# ===================================================================
#  Core Data Analysis & Quantitative Libraries
# ===================================================================
numpy~=2.3
pandas~=2.3

# Technical analysis indicators (forked & fixed version)
ft-pandas-ta~=0.3.15


# ===================================================================
#  Machine Learning & Training Libraries
# ===================================================================
tensorflow~=2.16
scikit-learn~=1.5
joblib~=1.4


# ===================================================================
#  API Clients & Database Connectors
# ===================================================================
python-binance~=1.0.29
python-telegram-bot~=22.1
pymongo~=4.13
sqlite-utils~=3.36


# ===================================================================
#  Utility Libraries
# ===================================================================
dateparser~=1.2
python-dotenv~=1.1.0
httpx~=0.27
setuptools~=69.5